resume_parser/requirements.txt
flask==2.3.3
werkzeug==2.3.7
pymupdf==1.23.9
spacy==3.7.2
python-docx==1.1.0
